requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
cachetools==5.3.1
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
Deprecated==1.2.14
Django==3.2.1
django-admin-model-list-order==0.0.4
django-appconf==1.0.5
django-cors-middleware==1.5.0
django-filter==23.2
django-imagekit==4.1.0
django-mdeditor==0.1.20
django-oauth-toolkit==2.3.0
django-oauth2-provider==0.2.6.1
django-prometheus==2.3.1
django-rest-swagger==2.2.0
django-simpleui==2023.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
dnspython==2.3.0
drf-yasg==1.21.6
eventlet==0.33.3
google-auth==2.20.0
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
jenkins==1.0.2
Jinja2==3.1.2
jwcrypto==1.5.0
kombu==5.3.1
kubernetes==26.1.0
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
pilkit==2.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
shortuuid==1.0.11
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.0
wrapt==1.15.0
